Overhaul _tt(string) functionality to work with wxWidgets
Fixes #166
