+++ src/ttTR.cpp
@@ -8,39 +8,23 @@
 
 #include "pch.h"
 
-#include <map>
-#include <optional>
-
-#if defined(_WX_DEFS_H_)
-    #include <wx/intl.h>
-#endif
-
 #include "ttTR.h"
-#include "ttcstr.h"
+
+ttTR::cmap tt_translations;
 
 namespace ttTR
 {
     ttlib::cstr trEmpty;
-
-    class cmap : public std::map<const std::string, ttlib::cstr>
-    {
-    public:
-        std::pair<bool, const ttlib::cstr&> getValue(const std::string& key) const
-        {
-            if (auto found = find(key); found != end())
-                return { true, found->second };
-            else
-                return { false, trEmpty };
-        }
-    };
 }  // namespace ttTR
 
-ttTR::cmap tt_translations;
+// This conditional is in case the module is being compiled directly instead of into a library (wxUiEditor does this)
 
-// We already have ttlib::emptystring, but it's a const and returning it causes a compiler error, so we create a
-// non-const here to return. Note that this does provide the option of setting the string to a specific value to
-// indicate an error, since it only gets returned if an empty string is passed, or in the case of translate(WORD id), it
-// is returned if the resource id doesn't exists.
+#if !defined(_WX_INTL_H_)
+
+// The _tt() macro can be used to load a string from the STRINGTABLE resource when compiled for Windows, or it can
+// specify a string when compiled for other platforms. If <wx/intl.h> has been loaded, then this function will never be
+// called. It exists for when translation is available for Windows builds, but not for other platforms and wxWidgets is
+// not available.
 
 const ttlib::cstr& ttlib::translate(const std::string& str)
 {
@@ -53,49 +37,21 @@ const ttlib::cstr& ttlib::translate(const std::string& str)
     {
         return value;
     }
-
     else
     {
-#if defined(_WX_DEFS_H_)
-        if (auto trans = wxTranslations::Get(); trans)
-        {
-            if (auto strTranslation = trans->GetTranslatedString(str); !strTranslation.empty())
-            {
-                if (auto [pair, success] = tt_translations.insert({ str, strTranslation->utf8_str().data() }, success))
-                {
-                    return pair->second;
-                }
-            }
-            else
-            {
-                // No translation available, make certain we don't try again.
-                if (auto [pair, success] = tt_translations.insert({ str, str }); success)
-                {
-                    return pair->second;
-                }
-            }
-        }
-        else
-        {
-            // No translation available, make certain we don't try again.
-            if (auto [pair, success] = tt_translations.insert({ str, str }); success)
-            {
-                return pair->second;
-            }
-        }
-#else   // !defined(_WX_DEFS_H_)
-        // Currently all that happens is the string is added without translation.
+        // Currently all that happens is the string is added without translation. Ideally, this should hook up directly
+        // to the xgettext translation system without requiring wxWidgets.
         if (auto [pair, success] = tt_translations.insert({ str, str }); success)
         {
             return pair->second;
         }
-#endif  // _WX_DEFS_H_
     }
 
-    // We only get here if the passed string ptr is null, or the unordered_map was unable to insert
-    // the string.
+    // We only get here if the passed string ptr is null, or the unordered_map was unable to insert the string.
     return ttTR::trEmpty;
 }
+
+#endif
 
 void ttlib::clearTranslations() noexcept
 {
